Fix depth yolo
